requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.1.3
django-celery-beat==2.4.0
django-filter==22.1
django-smtp-ssl==1.0
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flake8==5.0.4
gunicorn==20.1.0
kombu==5.2.4
mccabe==0.7.0
prompt-toolkit==3.0.32
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.6.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
vine==5.0.0
wcwidth==0.2.5
